test(integration): migrate remaining mobile-only tests to scenarioBuilder(mobileOnly)
Convert the rest of the legacy mobile-only tests off test: / twakePatrolTest
onto scenarioBuilder + mobileOnly: true (each legacy body wrapped in a scenario
driving the concrete mobile robots):
- chat_dm_leave (createDirectMessage unresolvable on web)
- chat_group display (test01) + copy (test04) — PullDownMenu / clipboard
- settings_recovery_key (Clipboard.getData)
- settings_contacts_visibility x3 (visibility backend absent locally)

Validated on web — all skipped, none failing:
- chat_group_test: 4 passed (reply/delete/edit/select), 3 skipped (display/copy/
  message-info), 0 failed
- chat_dm_leave / settings_recovery_key: skipped
- settings_contacts_visibility: 3 skipped

After this, the only remaining test: entry is chat_group test09 (message-info),
which is migrated in #3146; once that and these merge, the legacy test: /
twakePatrolTest signatures + RequiresLoginMixin can be removed (final cleanup).

Part of #3088 (Final).
